Possibility to add administrators to the db by running the SpringApplication
When the Spring application is run from the main method, the console will show three options: quit, run the application, add administrators.
For this to work, the "spring.jpa.hibernate.ddl-auto" property in application.properties should be set neither to 'create' nor to 'create-drod, because the application will quit after the administrators have been added to the database, causing them to be deleted.